--- src/main/java/com/project/coalba/domain/substituteReq/repository/SubstituteRepository.java
@@ -1,7 +1,0 @@
-package com.project.coalba.domain.substituteReq.repository;
-
-import com.project.coalba.domain.substituteReq.entity.SubstituteReq;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SubstituteRepository extends JpaRepository<SubstituteReq, Long>, SubstituteReqRepositoryCustom {
-}
