+++ src/main/java/com/project/coalba/domain/substituteReq/service/BossSubstituteReqService.java
@@ -5,7 +5,7 @@ import com.project.coalba.domain.schedule.entity.Schedule;
 import com.project.coalba.domain.substituteReq.dto.response.BothSubstituteReqResponse;
 import com.project.coalba.domain.substituteReq.dto.response.YearMonth;
 import com.project.coalba.domain.substituteReq.entity.SubstituteReq;
-import com.project.coalba.domain.substituteReq.repository.SubstituteRepository;
+import com.project.coalba.domain.substituteReq.repository.SubstituteReqRepository;
 import com.project.coalba.domain.substituteReq.repository.dto.BothSubstituteReqDto;
 import com.project.coalba.global.exception.BusinessException;
 import com.project.coalba.global.exception.ErrorCode;
@@ -21,18 +21,18 @@ import static java.util.stream.Collectors.groupingBy;
 @RequiredArgsConstructor
 @Service
 public class BossSubstituteReqService {
+    private final SubstituteReqRepository substituteReqRepository;
     private final ProfileUtil profileUtil;
-    private final SubstituteRepository substituteRepository;
 
     @Transactional(readOnly = true)
     public BothSubstituteReqDto getDetailSubstituteReq(Long substituteReqId) {
-        return substituteRepository.getSubstituteReq(substituteReqId);
+        return substituteReqRepository.getSubstituteReq(substituteReqId);
     }
 
     @Transactional(readOnly = true)
     public List<BothSubstituteReqResponse> getSubstituteReqs() {
         Boss currentBoss = profileUtil.getCurrentBoss();
-        List<BothSubstituteReqDto> bothSubstituteReqDtos = substituteRepository.getSubstituteReqs(currentBoss);
+        List<BothSubstituteReqDto> bothSubstituteReqDtos = substituteReqRepository.getSubstituteReqs(currentBoss);
 
         Map<YearMonth, List<BothSubstituteReqDto>> substituteReqMap = bothSubstituteReqDtos.stream()
                 .collect(groupingBy(bothSubstituteReqDto -> new YearMonth(bothSubstituteReqDto.getSubstituteReq().getCreatedDate())));
@@ -67,7 +67,7 @@ public class BossSubstituteReqService {
 
     @Transactional(readOnly = true)
     public SubstituteReq getSubstituteReqById(Long substituteReqId) {
-        return substituteRepository.findById(substituteReqId)
+        return substituteReqRepository.findById(substituteReqId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.SUBSTITUTEREQ_NOT_FOUND));
     }
 }
